+++ test/sql/type.sql
@@ -18,8 +18,8 @@ select (dlnewcopy(dlvalue('http://www.ljudmila.org'),1))->>'token' is not null;
 select dlpreviouscopy(dlvalue('http://www.ljudmila.org'),0);
 select (dlpreviouscopy(dlvalue('http://www.ljudmila.org'),1))->>'token' is not null;
 
-select dl_options(link_control=>'FILE');
-select dl_options(link_control=>'FILE',integrity=>'ALL',read_access=>'DB',write_access=>'BLOCKED');
-select * from dl_lco(dl_options(link_control=>'FILE',integrity=>'ALL',read_access=>'DB',write_access=>'BLOCKED'));
-select * from dl_lco(dl_options(link_control=>'FILE',integrity=>'ALL',recovery=>'YES',on_unlink=>'RESTORE'));
+select dl_lco(link_control=>'FILE');
+select dl_lco(link_control=>'FILE',integrity=>'ALL',read_access=>'DB',write_access=>'BLOCKED');
+select * from dl_link_control_options(dl_lco(link_control=>'FILE',integrity=>'ALL',read_access=>'DB',write_access=>'BLOCKED'));
+select * from dl_link_control_options(dl_lco(link_control=>'FILE',integrity=>'ALL',recovery=>'YES',on_unlink=>'RESTORE'));