+++ CHANGELOG.md
@@ -2,8 +2,10 @@ Version 0.2alpha
 -----------
 - datalink type move to pg_catalog
 - internally, datalink is now domain over jsonb
+- the head parameter to curl_get actually works now
 
 Version 0.1
 -----------
 - initial github version
 - moved SQL/MED standard functions such as dlvalue() to pg_catalog
+- removed curl_head() function. use curl_get()